refactor: share the value equality and visibility helpers
The adapter and the settings page each carried their own copy of
areValuesEqual and valueMatchesVisibilityRule, so the two could drift
apart. Moving them into a shared module keeps field visibility and
dirty-value checks behaving the same in both places.

Refs WOOPRD-3595
