+++ packages/js/settings-ui/src/settings-ui-page.tsx
@@ -43,6 +43,7 @@ import type {
 	SettingsValue,
 	SettingsValues,
 } from './types';
+import { areValuesEqual, valueMatchesVisibilityRule } from './values';
 
 type SaveNotice = {
 	status: 'success' | 'error';
@@ -69,19 +70,6 @@ const getInitialValues = ( schema: SettingsUISchema ): SettingsValues => {
 	} );
 
 	return values;
-};
-
-const areValuesEqual = ( a: SettingsValue, b: SettingsValue ) => {
-	if ( Array.isArray( a ) || Array.isArray( b ) ) {
-		return (
-			Array.isArray( a ) &&
-			Array.isArray( b ) &&
-			a.length === b.length &&
-			a.every( ( value, index ) => value === b[ index ] )
-		);
-	}
-
-	return a === b;
 };
 
 const getChangedValues = (
@@ -272,19 +260,6 @@ const GroupHeader = ( { group }: { group: SettingsUIGroup } ) => {
 				</div>
 			) : null }
 		</header>
-	);
-};
-
-const valueMatchesVisibilityRule = (
-	value: SettingsValue,
-	expected: SettingsValue | SettingsValue[] | undefined
-) => {
-	const expectedValues = Array.isArray( expected )
-		? expected
-		: [ expected ?? true ];
-
-	return expectedValues.some( ( expectedValue ) =>
-		areValuesEqual( value, expectedValue )
 	);
 };
